Fixed test_obdsim.py failure
The async connection test was failing, due to the first RPM query returning null.
This was due to 36784bffc2ca50bad7b0f7941189eda8a9865467, which has python-OBD
learning the number of messages to expect (rather than a dumb ECU count, which
may not be true for all commands). In this case, "fast" mode will not attach
a postfix message count, and will opt to incur the ELM's timeout on the first
query only, in order to learn the number of replies. This extra firt-message
initialization cost is more correct, but resulted in a timeout needing to be
tuned in the unit test.