Add support of custom headers
Allow usage of the optional `header` argument, which tells python-OBD
to use a custom header when querying the command. If not set, python-OBD
assumes that the default 7E0 header is needed for querying the command.
The switch between default and custom header (and vice versa) is
automatically done by python-OBD.

Also restore the default header when closing the connection.