check_voltage argument on OBD and Async
Added check_voltage optional argument that is `True` by default and when
set to `False` disables the detection of the car supply voltage on OBDII
port (which should be about 12V). This control assumes that, if the
voltage is lower than 6V, the OBDII port is disconnected from the car.
If the option is enabled, it adds the `OBDStatus.OBD_CONNECTED` status,
which is set when enough voltage is returned (socket connected to the
car) but the ignition is off (no communication with the vehicle).
Setting the option to `False` should be needed when the adapter does not
support the voltage pin or more generally when the hardware provides
unreliable results, or if the pin reads the switched ignition voltage
rather than the battery positive (this depends on the car).